try to fixed payment bug

+++ app/controller/orders.py
@@ -48,12 +48,14 @@ def create_order():
                         "message": response_content["msg"]
                     }
                 }
-                order_status = get_db("INSERT INTO orders (orderId, paymentStatus, orderName, orderEmail, orderPhone, orderItem, orderUser, date, time, price, attraction, address, image)VALUES(%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)", [order_number, response_content['status'], contact['name'], contact['email'], contact['phone'], attraction['id'], email, booking['date'], booking['time'],data['order']['price'], attraction['name'], attraction['address'], attraction['image']], 'none')
+                value = [order_number, response_content['status'], contact['name'], contact['email'], contact['phone'], attraction['id'], email, booking['date'], booking['time'],data['order']['price'], attraction['name'], attraction['address'], attraction['image']]
+                order_status = get_db("INSERT INTO orders (orderId, paymentStatus, orderName, orderEmail, orderPhone, orderItem, orderUser, date, time, price, attraction, address, image)VALUES(%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)", value, 'none')
                 update_status = get_db("UPDATE booking SET paymentStatus=0 WHERE bookingItem=%s", [attraction['id']], 'none')
                 undelete_items = get_db("SELECT bookingItem FROM booking WHERE bookingUser=%s AND paymentStatus=1", [email], 'all')
-                for item in undelete_items:
-                    id = item['bookingItem']
-                    delete_item = get_db("DELETE FROM booking WHERE bookingItem=%s", [id], 'none')
+                if(undelete_items != []):
+                    for item in undelete_items:
+                        id = item['bookingItem']
+                        delete_item = get_db("DELETE FROM booking WHERE bookingItem=%s", [id], 'none')
                 response = make_response({"data": response_data}, 200)
             else:
                 response = make_response({"error": True, "message": "Failed to pay", "order_number": order_number}, 400)
